+++ app/models/landscaping_job.rb
@@ -1,5 +1,2 @@
 class LandscapingJob < ApplicationRecord
-    def index
-        @LandscapingJobs = LandscapingJob.all
-      end
 end